moving ftxs classes to ledger lib (wip)
