+++ components/ledger/ledger-utxo-flow/src/test/kotlin/net/corda/ledger/utxo/flow/impl/transaction/filtered/factory/UtxoFilteredTransactionFactoryImplTest.kt
@@ -6,8 +6,8 @@ import net.corda.ledger.common.testkit.publicKeyExample
 import net.corda.ledger.lib.utxo.flow.impl.timewindow.TimeWindowBetweenImpl
 import net.corda.ledger.lib.utxo.flow.impl.transaction.UtxoSignedTransactionInternal
 import net.corda.ledger.utxo.data.transaction.UtxoComponentGroup
-import net.corda.ledger.utxo.flow.impl.transaction.filtered.UtxoFilteredTransactionBuilderImpl
-import net.corda.ledger.utxo.flow.impl.transaction.filtered.UtxoFilteredTransactionBuilderInternal
+import net.corda.ledger.lib.utxo.flow.impl.transaction.UtxoFilteredTransactionBuilderImpl
+import net.corda.ledger.lib.utxo.flow.impl.transaction.UtxoFilteredTransactionBuilderInternal
 import net.corda.ledger.utxo.test.UtxoLedgerTest
 import net.corda.ledger.utxo.testkit.notaryX500Name
 import net.corda.ledger.utxo.testkit.utxoTimeWindowExample
